Consolidate test configuration into rails_helper
We really don’t have very much configuration for
the test suite and I think it’s easier to find
things if we have a single place for it all in
rails_helper.
