+++ db/seeds.rb
@@ -1,7 +1,4 @@
-if User.where(name: "Test user").present?
-  puts "Skipping because user already exists"
-  return
-end
+return if User.where(name: "Test user").present?
 
 gds_organisation_id = "af07d5a5-df63-4ddc-9383-6a666845ebe9"
