+++ spec/models/user_spec.rb
@@ -1,5 +1,3 @@
-require "gds-sso/lint/user_spec"
-
 RSpec.describe User, type: :model do
   it_behaves_like "a gds-sso user class"
 end
