+++ spec/requests/get_subscriber_list_spec.rb
@@ -182,6 +182,6 @@ RSpec.describe "Getting a subscriber list", type: :request do
   end
 
   def get_subscriber_list(query_payload)
-    get "/subscriber-lists", params: query_payload, headers: json_headers
+    get "/subscriber-lists", params: query_payload, headers: JSON_HEADERS
   end
 end
