+++ spec/requests/send_notification_spec.rb
@@ -127,7 +127,7 @@ RSpec.describe "Sending a notification", type: :request do
         tags: tags,
       }.merge(options))
 
-      post "/notifications", params: request_body, headers: json_headers
+      post "/notifications", params: request_body, headers: JSON_HEADERS
     end
   end
 
@@ -161,7 +161,7 @@ RSpec.describe "Sending a notification", type: :request do
       allow(NotificationWorker).to receive(:perform_async)
       post "/notifications",
         params: request_params,
-        headers: json_headers.merge("HTTP_GOVUK_REQUEST_ID" => "request_id")
+        headers: JSON_HEADERS.merge("HTTP_GOVUK_REQUEST_ID" => "request_id")
     end
 
     it "creates a Notification" do
