+++ spec/services/email_sender_service_spec.rb
@@ -1,5 +1,3 @@
-require "notifications/client"
-
 RSpec.describe EmailSenderService do
   it "sends an email to the override email address" do
     config = { provider: "NOTIFY", email_address_override: "override@example.com" }
