--- spec/spec_helper.rb
@@ -1,29 +1,0 @@
-# This file was generated by the `rails generate rspec:install` command. Conventionally, all
-# specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
-# The generated `.rspec` file contains `--require spec_helper` which will cause this
-# file to always be loaded, without a need to explicitly require it in any files.
-#
-# Given that it is always loaded, you are encouraged to keep this file as
-# light-weight as possible. Requiring heavyweight dependencies from this file
-# will add to the boot time of your test suite on EVERY test run, even for an
-# individual file that may not need all of that loaded. Instead, make a
-# separate helper file that requires this one and then use it only in the specs
-# that actually need it.
-#
-# The `.rspec` file also contains a few flags that are not defaults but that
-# users commonly want.
-#
-# See http://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
-
-$LOAD_PATH << File.join(File.dirname(__FILE__), '..')
-
-require 'webmock/rspec'
-WebMock.disable_net_connect!(allow_localhost: true)
-
-require 'base64'
-HTTP_AUTH_CREDENTIALS = Base64.strict_encode64('test@example.com:test123')
-
-RSpec.configure do |config|
-  config.disable_monkey_patching!
-  config.expose_dsl_globally = false
-end
