--- spec/support/helpers/authentication_helper.rb
@@ -1,17 +1,0 @@
-module AuthenticationHelper
-  module ControllerMixin
-    def login_as(user)
-      request.env["warden"] = double(
-        authenticate!: true,
-        authenticated?: true,
-        user: user
-      )
-    end
-  end
-
-  module RequestMixin
-    def login_as(user)
-      GDS::SSO.test_user = user
-    end
-  end
-end
