--- spec/support/helpers/json_request_helpers.rb
@@ -1,8 +1,0 @@
-module JSONRequestHelpers
-  def json_headers
-    {
-      'CONTENT_TYPE' => "application/json",
-      'ACCEPT' => 'application/json'
-    }
-  end
-end
